Bug 512328 - Update liboggplay to rev 404316e595. r=doublec
UltraBlame original commit: 314c78cf870089e5134db3459077ab815271f61a

+++ media/liboggplay/src/liboggplay/oggplay_callback.h
@@ -57,7 +57,7 @@ oggplay_process_leftover_packet(OggPlay *me);
 
 
 OggPlayDecode *
-oggplay_initialise_decoder(OggPlay *me, int content_type, int serialno);
+oggplay_initialise_decoder(OggPlay *me, int content_type, long serialno);
 
 int
 oggplay_callback_info_prepare(OggPlay *me, OggPlayCallbackInfo ***info);
